requirements.txt
cffi==1.15.1
cryptography==41.0.2
Pillow==10.0.0
pycparser==2.21
pycryptodome==3.18.0
